fix(config): Move core config to clouddriver.config package
