+++ clouddriver-core/src/main/groovy/com/netflix/spinnaker/clouddriver/config/LocalJobConfig.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.config
+package com.netflix.spinnaker.clouddriver.config
 
 import com.netflix.spinnaker.clouddriver.jobs.JobExecutor
 import com.netflix.spinnaker.clouddriver.jobs.local.JobExecutorLocal
