+++ clouddriver-web/src/main/groovy/com/netflix/spinnaker/clouddriver/Main.groovy
@@ -37,6 +37,7 @@ import java.security.Security
 ])
 @ComponentScan([
   'com.netflix.spinnaker.config',
+  'com.netflix.spinnaker.clouddriver.config'
 ])
 @EnableAutoConfiguration(exclude = [
     BatchAutoConfiguration,
